code refactoring and code clean
