+++ src/authSession/authSession.ts
