+++ test/inboxLogic.test.ts
@@ -1,7 +1,6 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 import solidNamespace from "solid-namespace";
 import * as rdf from "rdflib";
-import fetchMock from "jest-fetch-mock";
 import { UpdateManager } from "rdflib";
 import { SolidNamespace } from "../src/index";
 import { ProfileLogic } from "../src/profile/ProfileLogic";
